Added TakeDamage Function
