+++ Source/Redemption/Public/RedemptionPlayer.h
@@ -4,6 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Character.h"
+#include "Components/WidgetComponent.h"
 #include "RedemptionPlayer.generated.h"
 
 class USpringArmComponent;
@@ -12,6 +13,7 @@ class UInputMappingContext;
 class UInputAction;
 struct FInputActionValue;
 class UPowerUpComponent;
+class UHealthComponent;
 
 UCLASS()
 class REDEMPTION_API ARedemptionPlayer : public ACharacter
@@ -40,6 +42,8 @@ protected:
 	UPROPERTY(EditDefaultsOnly, Category = "Input")
 	UInputAction* LookAction;
 
+	UHealthComponent* HealthComp;
+
 public:
 	ARedemptionPlayer();
 
@@ -54,16 +58,18 @@ public:
 
 	virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
 
-	float GetHealth() const;
-	float GetMaxHealth() const;
-	void SetHealth(float const NewHealth);
+	//float GetHealth() const;
+	//float GetMaxHealth() const;
+	//void SetHealth(float const NewHealth);
+	//
+	//void TakeDamage();
 
 public:
 	FORCEINLINE class USpringArmComponent* GetCameraBoom() const { return CameraBoom; }
 	FORCEINLINE class UCameraComponent* GetFollowCamera() const { return FollowCamera; }
 
 private:
-	class UWidgetComponent* WidgetComponent;
-	float const MaxHealth{ 100.f };
-	float Health;
+	//class UWidgetComponent* WidgetComponent;
+	//float const MaxHealth{ 100.f };
+	//float Health;
 };
